add check in mint() to prevent sending token to the token contract itself.
The same was already done for transfer(). Due to the internal design of ERC20.sol of openzeppelin-solidity@2.1.3, it is not possible for transferFrom(). Instead of updating to a newer version of openzeppelin-solidity, we decided to leave transferFrom() as is.

Issue found in a Smart Contract Audit by SOMISH Blockchain Labs (@somish) (Minor 3. Missing check in mint(),transferFrom() functions.) Thanks!